backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.4
django-otp==1.6.3
djangorestframework==3.16.1
kombu==5.6.2
packaging==26.1
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
qrcode==8.2
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
